requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
haystack-ai==2.5.1
haystack-experimental==0.1.1
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.5.0
narwhals==1.8.2
networkx==3.3
numpy==1.26.4
openai==1.47.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
posthog==3.6.6
protobuf==5.28.2
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pypdf==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
qdrant-client==1.11.2
qdrant-haystack==6.0.0
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
